domain/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sensor-domain"
version = "0.1.0"
description = "Domain models for sensor failure detection system"
requires-python = ">=3.10"
dependencies = [
    "tortoise-orm>=0.25.0",
]

[tool.setuptools.packages.find]
# Tell setuptools to search in the current directory
where = ["."]
# Include all packages
include = ["domain*"]
# No exclusions
exclude = []